requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.6.0
torchvision==0.7.0
asgiref==3.2.10
Django==3.0.8
djangorestframework==3.11.0
Pillow==7.2.0
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
